Initial Commit
This is a maestro test of clicking each element of the following list in the X app. Maestro is used to access each element of a list with an unknown number of elements, which we usually find in social media apps, in the case of lists of following/followers, and can be generally applied to any list.